+++ UnityProject/Assets/Scripts/Items/IDCard.cs
@@ -10,7 +10,7 @@ using WebSocketSharp;
 /// <summary>
 ///     ID card properties
 /// </summary>
-public class IDCard : NetworkBehaviour, IServerInventoryMove, IServerSpawn, IInteractable<HandActivate>
+public class IDCard : NetworkBehaviour, IServerInventoryMove, IServerSpawn, IInteractable<HandActivate>, IExaminable
 {
 
 	[Tooltip("Sprite to use when the card is a normal card")]
@@ -57,6 +57,7 @@ public class IDCard : NetworkBehaviour, IServerInventoryMove, IServerSpawn, IInt
 	[SyncVar(hook = nameof(SyncName))]
 	private string registeredName;
 
+	public int[] currencies = new int[(int)CurrencyType.Total];
 
 	//The actual list of access allowed set via the server and synced to all clients
 	private SyncListInt accessSyncList = new SyncListInt();
@@ -273,4 +274,17 @@ public class IDCard : NetworkBehaviour, IServerInventoryMove, IServerSpawn, IInt
 	{
 		SyncName(registeredName, newName);
 	}
+
+	public string Examine(Vector3 pos)
+	{
+		return $"The account linked to the ID belongs to '{registeredName}' and reports a balance of {currencies[(int)CurrencyType.Credits]} cr. " +
+		       $"The mining budget reports a balance of {currencies[(int)CurrencyType.LaborPoints]} points.";
+	}
+}
+
+public enum CurrencyType
+{
+	Credits,
+	LaborPoints,
+	Total
 }